Integrate fan service within joint_state.
